minor fix: typedef of VA makes compilation on 32-bit machines fail (#174)
Closes Issue #160 

* change typedef VA to uint64_t

+++ mcsema/Arch/Arch.h
@@ -116,7 +116,7 @@ class MCInst;
 
 }  // namespace llvm
 
-typedef uintptr_t VA;
+typedef uint64_t VA;
 
 enum SystemArchType {
   _X86_,
